requirements.txt
pydantic
pydantic-settings
pydantic-ai
gradio
python-dotenv
ddgs
httpx
